code fix for big decimal to double

+++ authentication/authentication-common/src/main/java/io/mosip/authentication/common/service/entity/PartnerCurrentBalance.java
@@ -1,6 +1,5 @@
 package io.mosip.authentication.common.service.entity;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import javax.persistence.Column;
@@ -30,7 +29,7 @@ public class PartnerCurrentBalance {
 
 	@NotNull
 	@Column(name = "balance", precision = 19, scale = 2)
-	private BigDecimal balance;
+	private Double balance;
 
 	@NotNull
 	@Column(name = "cr_by")
